requirements.txt
aniso8601==9.0.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cfgv==3.3.1
chardet==3.0.4
circuitbreaker==1.3.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.4.2
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-Log-Request-ID==0.10.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
flickrapi==2.4.0
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
identify==2.4.9
idna==2.8
importlib-metadata==1.5.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
neat-python==0.92
netifaces==0.10.4
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pbr==5.8.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.0
pre-commit==2.17.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==4.2.0
reportlab==3.5.34
requests==2.25.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
toml==0.10.2
typing-extensions==4.1.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.1
wadllib==1.3.3
Werkzeug==2.0.2
wrapt==1.14.0
WTForms==3.0.1
xkit==0.0.0
zipp==1.0.0
